Changed from tri_idxs from size_t to int and fixed AVX for x86; changed loading method for aarch64 to use consecutive memory addresses; moved printing of the BVH tree before rendering

+++ src/ADS/BVH/BVHVec/bvh_vec.h
@@ -59,7 +59,7 @@ class BvhVec : public Ads {
     void reset_stats() const;
 
     std::vector<BvhNode> nodes_;  // root node at idx 0, child nodes are parent node_idx*2+1 and node_idx*2+2
-    std::vector<size_t> triangle_indices_;
+    std::vector<int> triangle_indices_;
 
     // Config variables
     size_t max_triangles_per_BB_;
